Move function docs to another file
The explanation of why functions and procedural abstraction are useful
was a bit too long for the project spec, so put it in `functions.md`
instead.

The fact that the markdown files explaining the project individual parts
were actually numbered made it harder to rearrange their order in the
spec, since we'd have to rename the files as well. This would cause git
shenanigans, so change the nomenclature for part specs so that their
names are descriptive of what they really are.

Start writing explanations of basic C++ syntax, and other common
mistakes.